Move lj_auditlog.c out of Snabb and into RaptorJIT
Git had misplaced this file in src/ instead of lib/luajit/src/
